pyproject.toml
[project]
name = "aklatan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.42",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]

[tool.pytest.ini_options]
pythonpath = "src"

